+++ src/main/scala/ch/openolitor/core/data/parsers/AbotypParser.scala
@@ -66,7 +66,7 @@ object AbotypParser extends EntityParser {
           case fristeinheitPattern(wert, "M") => Frist(wert.toInt, Monatsfrist)
         },
         anzahlAbwesenheiten = row.value[Option[Int]](indexAnzahlAbwesenheiten),
-        anzahlEinsaetze = row.value[Option[Int]](indexAnzahlEinsaetze),
+        anzahlEinsaetze = row.value[Option[BigDecimal]](indexAnzahlEinsaetze),
         farbCode = row.value[String](indexFarbCode),
         zielpreis = row.value[Option[BigDecimal]](indexZielpreis),
         guthabenMindestbestand = row.value[Int](indexGuthabenMindestbestand),
