+++ src/main/scala/ch/openolitor/core/db/evolution/scripts/v1/V1Scripts.scala
@@ -169,7 +169,7 @@ object V1Scripts {
         vertragslaufzeit varchar(50),
         kuendigungsfrist varchar(50),
         anzahl_abwesenheiten int,
-        anzahl_einsaetze int,
+        anzahl_einsaetze DECIMAL(5,2),
         farb_code varchar(20),
         zielpreis DECIMAL(7,2),
         guthaben_mindestbestand int,
